requirements.txt
requests
beautifulsoup4
openpyxl
selenium
meraki
python-dotenv